fix row Major bug

+++ utils/cublas_lay_egg.h
@@ -2,11 +2,11 @@
 #include <cuda_runtime.h>
 #include "cublas_utils.h"
 
-void verify(int M, int N, half* A, int lda, half* B, int ldb, double rerror);
+void verify(int M, int N, half* A, int lda, float* B, int ldb, double rerror);
 void verify_blas(int M, int N, int K,
                  const half* Ah, int lda,
                  const half* Bh, int ldb,
                  half* Ch, int ldc,
                  const half alpha,
                  const half beta,
-                 half* Dh);
+                 float* Dh);
